Engineering Task #3412 (7 hours): 1. Refactored generic SDT WSDLs and XSDs to remove references to specific Case Management system like MCOL. 2. Adjusted wsdl to source code generation target to generate sources in 'work' directory. This is to prevent losing SVN meta-data for existing src-gen/main directory.
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@16985 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ domain/src/main/java/uk/gov/moj/sdt/domain/BulkSubmission.java
@@ -79,7 +79,7 @@ public class BulkSubmission extends AbstractDomainObject implements IBulkSubmiss
     /**
      * The total number of requests expected to be contained within the Bulk Request.
      */
-    private int numberOfRequest;
+    private long numberOfRequest;
 
     /**
      * The status of the Bulk Request - one of "Uploaded", "Failed", "Validated", or "Completed"
@@ -181,13 +181,13 @@ public class BulkSubmission extends AbstractDomainObject implements IBulkSubmiss
     }
 
     @Override
-    public int getNumberOfRequest ()
+    public long getNumberOfRequest ()
     {
         return numberOfRequest;
     }
 
     @Override
-    public void setNumberOfRequest (final int numberOfRequest)
+    public void setNumberOfRequest (final long numberOfRequest)
     {
         this.numberOfRequest = numberOfRequest;
     }
